refactor: move logChatEvent to a helper rather than thunk
